+++ QUANTAXIS/QACmd/__init__.py
@@ -211,7 +211,7 @@ class CLI(cmd.Cmd):
                 # QA_SU_save_etf_day('tdx')
                 # QA_SU_save_etf_min('tdx')
                 QA_SU_save_stock_list('tdx')
-                # QA_SU_save_stock_block('tdx')
+                QA_SU_save_stock_block('tdx')
                 # QA_SU_save_stock_info('tdx')
             elif len(arg) == 1 and arg[0] == 'day':
                 if QA_Setting().client.quantaxis.user_list.find({'username': 'admin'}).count() == 0: